+++ common/src/main/java/fr/loudo/narrativecraft/client/editors/cutscene/ClientCutsceneMakerEditorMaker.java
@@ -223,6 +223,8 @@ public class ClientCutsceneMakerEditorMaker implements EditorMaker {
         if (cutscene.getEditorLayers() != null) {
             editorLayers.addAll(cutscene.getEditorLayers());
         }
+        totalTick = cutscene.getMaxTick();
+        playback.setTotalTick(totalTick);
     }
 
     public void addLayer(ICutsceneLayer layer) {