+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 - Some crash case while trying to show a character skin
 - Crash when opening the global dialog editor with no target character
+- Prevent saving an animation while still recording
 
 ## [2.1.0]