+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [2.1.1]
 
+### Features
+
+- Improved story manager UI
+
 ### Fixes
 - Some crash case while trying to show a character skin
 - Crash when opening the global dialog editor with no target character