+++ CHANGELOG.md
@@ -2,5 +2,8 @@
 
 ## [2.1.2]
 
+### Fixes
+
 - Ride action not working correctly on some furniture mod
 - Entity custom name visible if not character in playback
+- Cutscene max tick conflict if removing an animation/subscene