+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - Ride action not working correctly on some furniture mod
 - Entity custom name visible if not character in playback
 - Cutscene max tick conflict if removing an animation/subscene
+- Revert fix on story (break choices appear)