+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 
 - World locked if leaving the world while a playback is playing
 - Character placement on ground in camera angle not accurate
+- Log in chat of executed command with command tag