+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [2.1.2]
+## [2.1.3]
 
 ### Fixes
 
@@ -8,3 +8,4 @@
 - Entity custom name visible if not character in playback
 - Cutscene max tick conflict if removing an animation/subscene
 - Revert fix on story (break choices appear)
+- World locked if leaving the world while a playback is playing